Move Character interface to data namespace
