+++ deploy/lib/control/Player.class.php
@@ -1,5 +1,4 @@
 <?php
-require_once(CORE . "control/Character.php");
 require_once(CORE . "control/lib_status.php");
 require_once(CORE . "control/lib_player.php");
 
@@ -8,6 +7,7 @@ use NinjaWars\core\data\ClanFactory;
 use NinjaWars\core\data\Clan;
 use NinjaWars\core\data\PlayerDAO;
 use NinjaWars\core\data\PlayerVO;
+use NinjaWars\core\data\Character;
 
 /* Ninja (actually character) behavior object.
  *
